TimesHackathon/engineers/requirements.txt
gdown==5.2.0
pandas==2.2.2
python-dotenv==1.0.1
sqlalchemy==2.0.32
psycopg2-binary
streamlit==1.37.1
rapidfuzz